+++ CHANGELOG.md
@@ -31,6 +31,8 @@ All notable changes to this project will be documented in this file.
 - Fix a longstanding problem of including empty `categories`, `shortNames` and `additionalPrinterColumns` in the CRDs,
   which could cause problems with GitOps tools (e.g. ArgoCD) reporting a diff in the custom resources.
   See [our internal issue](https://github.com/stackabletech/hdfs-operator/issues/626) and [the fix](https://github.com/kube-rs/kube/pull/2042) for details ([#741]).
+- The operator now watches all resources that it creates and early-exits the reconcile action when the
+  cluster is marked for deletion ([#754]).
 
 [#726]: https://github.com/stackabletech/hive-operator/pull/726
 [#731]: https://github.com/stackabletech/hive-operator/pull/731
@@ -39,6 +41,7 @@ All notable changes to this project will be documented in this file.
 [#737]: https://github.com/stackabletech/hive-operator/pull/737
 [#741]: https://github.com/stackabletech/hive-operator/pull/741
 [#748]: https://github.com/stackabletech/hive-operator/pull/748
+[#754]: https://github.com/stackabletech/hive-operator/pull/754
 
 ## [26.7.0] - 2026-07-21